+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue [#117](https://github.com/m-marini/wheellino/issues/117): Add Lidar sensors
 
+### Changed
+
+- Issue [#120](https://github.com/m-marini/wheellino/issues/120): Change motor controller
+
 ## [0.9.0] 2025-10-06
 
 ### Added